fix: incorrect check for existing socket clients
Fixes #61. Check for existing socket clients didn't work correctly
because wss.clients is a Set, not an Array. Change the "length" check to
check property "size" instead.

Also, change the expected amount of connected clients from 0 to 1. The
check is inside "connected" event handler so there should be at least
the client whose connection invoked said handler.

This is the same fix as in commit c4e37db325, but to the v3.0.0 version
of the library.